+++ templates/gpxcontent.php
@@ -142,6 +142,17 @@ if (count($_['dirs']) === 0){
                 <label for="displayclusters"><i class="fa fa-map-marker" aria-hidden="true" style="color:blue;"></i>
                 <?php p($l->t('Display markers'));?></label>
             </div>
+            <div title="<?php
+p($l->t('Open popups of pictures markers'));
+echo "\n\n";
+p($l->t('Pictures of current folder are displayed only if you have Python PIL installed'));
+echo "\n\n";
+p($l->t('Only pictures with EXIF geolocation data are displayed')); ?>">
+                <input id="showpicscheck" type="checkbox" checked="checked">
+                <label for="showpicscheck">
+                <i class="fa fa-map-marker" aria-hidden="true" style="color:red;"></i>
+                <?php p($l->t('Show pictures')); ?></label>
+            </div>
             <div title="<?php p($l->t('Open info popup when a track is drawn')); ?>">
                 <input id="openpopupcheck" type="checkbox" checked="checked">
                 <label for="openpopupcheck"><i class="fa fa-comment-o" aria-hidden="true"></i>